+++ src/hyperbase/__init__.py
@@ -1,6 +1,8 @@
 from hyperbase.hyperedge import hedge
+from hyperbase.parsers import get_parser
 
 
 __all__ = [
-    'hedge'
+    'hedge',
+    'get_parser',
 ]